Use nicehash api to get algo ids
Previously we had a hard coded list of algorithms that we used to
determine the appropriate nicehash id for our api calls. This list was
out of date causing us to miss some coins. This code uses the nicehash
api to retrieve these ids to make sure they are always up to date.